Pre-Release 3.0.6: Updated To Support 4.89 HFW
DISCLAIMER: There will be XMB artifacts and alignment issues on some items and menus. This is due to the DEX version of explore_plugin.sprx being needed for package manager to work properly.

It is advised to install the HAN_489_Fix_272p_480p_Resolutions_signed.pkg which is the same fix that PS3HEN uses, from sandungas.